fix1

+++ todo_service/service.js
@@ -1,4 +1,51 @@
-const { TodoModel } = require("./model");
+const {Axios } = require("axios");
+const { GATEWAY_URI } = require("./config");
+const service = require("./service");
+const TodoModel = require("./model");
+
+const axios = new Axios({
+    timeout: 3000,
+});
+
+async function getUser(userId) {
+    const response = await axios.get(`${GATEWAY_URI}/user/findById/?id=${userId}`);
+    if (response.status === 200) {
+        return response.data.user;
+    }
+
+    // Failed
+    return undefined;
+}
+
+async function increaseTodo(userId) {
+    const user = await getUser(userId);
+    if(!user){
+        return undefined;
+    }
+
+    const response = await axios.get(`${GATEWAY_URI}/user/setTodos/?number=${user.numberTodo + 1}`);
+    if (response.status === 200) {
+        return true;
+    }
+
+    // Failed
+    return undefined;
+}
+
+async function decreaseTodo(userId) {
+    const user = await getUser(userId);
+    if(!user){
+        return undefined;
+    }
+
+    const response = await axios.get(`${GATEWAY_URI}/user/setTodos/?number=${user.numberTodo - 1}`);
+    if (response.status === 200) {
+        return true;
+    }
+
+    // Failed
+    return undefined;
+}
 
 //Dealing with data base operations
 class TodoRepository {
@@ -7,9 +54,18 @@ class TodoRepository {
      * @param {string} userId
      * @returns list todo
      */
-    async GetAll(userId) {
+    async GetByUserId(userId) {
         try {
-            const todos = await TodoModel.getAll(userId);
+            const todos = await TodoModel.getByUserId(userId);
+            return todos;
+        } catch (err) {
+            throw error;
+        }
+    }
+
+    async GetAll() {
+        try {
+            const todos = await TodoModel.getAll();
             return todos;
         } catch (err) {
             throw error;
@@ -23,7 +79,7 @@ class TodoRepository {
      */
     async GetOne(todoId) {
         try {
-            const todos = await TodoModel.getOne(todoId);
+            const todos = await TodoModel.getById(todoId);
             return todos;
         } catch (err) {
             throw error;
@@ -46,21 +102,34 @@ class TodoRepository {
 
     async Create({ name, userId }) {
         try {
-            const isSuccess = await TodoModel.create(name, userId);
-            return isSuccess;
+            if(await increaseTodo(userId)){
+                await TodoModel.create(name, userId);
+                return true;
+            }
+            else{
+                console.log("failed when call");
+            }
+            
+            return undefined;
         } catch (error) {
+            console.log(error);
             throw error;
         }
     }
 
     async Delete(todoId) {
         try {
-            const isSuccess = await TodoModel.delete(todoId);
-            return isSuccess;
+            const todo = await TodoModel.getById(todoId);
+            if(todo && await decreaseTodo(todo.userId)){
+                await TodoModel.delete(todoId);
+                return true;
+            }
+            
+            return undefined;
         } catch (error) {
             throw error;
         }
     }
 }
 
-module.exports = TodoRepository;
+module.exports = new TodoRepository();
